collector/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
prometheus-client==0.19.0
pyyaml==6.0.1

# Testing (not installed in production image)
# pip install pytest pytest-asyncio
